Moved CalendarsToPrint to component file, merged 2 checkboxes, cleanup
