+++ frontend/src/components/styles/CalendarsToPrint.css
@@ -2,12 +2,6 @@
     display: none;
 }
 
-.center-with-margin {
-    margin-top: 15px;
-    display: flex;
-    justify-content: center;
-}
-
 .hide-navigation .rbc-btn-group {
     display: none !important;
 }
